requirements.txt
fastapi
uvicorn
streamlit
requests
python-dotenv

langchain
langchain-community
langchain-google-genai
langchain-classic

duckduckgo-search
wikipedia
arxiv
pymed
tavily-python

beautifulsoup4
faiss-cpu
chromadb
